+++ lib/service/Auth.dart
@@ -611,7 +611,7 @@ class Auth with Service implements NotificationsListener {
 
       try {
         Response response = await Network().delete(url, headers: {'Content-Type': 'application/json'}, auth: NetworkAuth.User);
-        if(response?.statusCode == 200) {
+        if((response != null) && (200 <= response.statusCode) && (response.statusCode <= 205)) { // returns 202 Accepted
           _applyUserPiiData(null, null);
           return true;
         }