+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Introduced Family members panel [#537](https://github.com/rokwire/safer-illinois-app/issues/537).
 
 ## [2.9.13] - 2021-02-22
 ### Changed
 - Updated Positive IP & NIP step & explanation strings [#529](https://github.com/rokwire/safer-illinois-app/issues/529).
 - Make POSITIVE-NIP COVID-19 PCR test result like the regular "PCR.negative" test result [#532](https://github.com/rokwire/safer-illinois-app/issues/532).
-- Introduced Approve Family members panel [#534](https://github.com/rokwire/safer-illinois-app/issues/534).
+- Introduced Approve Family member panel [#534](https://github.com/rokwire/safer-illinois-app/issues/534).
 
 ### Added
 - Added camera usage mention in disclosure screen [#530](https://github.com/rokwire/safer-illinois-app/issues/530).