+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make the points from the first section from OnBoardingHealthDisclosurePanel panel bold [#489](https://github.com/rokwire/safer-illinois-app/issues/489),
 - Updated Positive IP & NIP step & explanation strings [#529](https://github.com/rokwire/safer-illinois-app/issues/529).
 - Added camera usage mention in disclosure screen [#530](https://github.com/rokwire/safer-illinois-app/issues/530)
+- Updated PCR.positive-NIP test rule [#532](https://github.com/rokwire/safer-illinois-app/issues/532)
 
 ### Added
 - App Up time to exposure logs.