+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fixed "Remove My Information" processing [#547](https://github.com/rokwire/safer-illinois-app/issues/547).
 
 ## [2.9.15] - 2021-02-26
 ### Added